Adding speex, changed filename;
Version 1.0.2
The changes in this release are as follows:
This update adds soundcard support for Solaris and the BSDs as well as
minor bugfixes and a documentation update.

Version 1.1
The changes in this release are as follows:
This release adds many new features, including a denoiser that removes
most of the background noise and can be used before encoding
(available as --denoise in speexenc), adaptive gain control which
adjusts the volume to a constant level (available as --agc in
speexenc), echo cancellation (experimental, only available from
libspeex), a jitter buffer which handles packet buffering for VoIP
applications (experimental, only available from libspeex), and an
improved narrowband mode at 4.8 kbps (available only from libspeex
for applications that don't require compatibility).

Version: 1.1.1
The changes in this release are as follows:
This release adds a partial fixed-point port which can be enabled
using the --enable-fixed-point option at configure time. Not all
floating-point operations have been converted yet, but all the code
should work.