New upstream release;
2.6.11
The changes in this release are as follows:
This release fixes the mirror for empty files, a rare coredump in HTTP
0.9 handling which occurred when the first line of a reply was empty,
compilation problems for compilers other than gcc, and translations.

2.6.12
The changes in this release are as follows:
This release fixes put over HTTPS, spinning in bg mode when an
external program was still running at the time of moving to the
background, and compilation with the IRIX MIPSPro compiler.