New upstream release, finally adding the full xmltv variant to 10.3;
0.5.31
The changes in this release are as follows:
Problems have been fixed with summer time changeover in many of the
grabbers. There is a new tv_grab_na_dd grabber for North America.
There are some other fixes to grabbers and to tv_check.

0.5.30
The changes in this release are as follows:
This release adds new grabbers for Japan, Germany, Sweden, and France.
The UK grabber now does radio channels. Some bugfixes have been made.

0.5.29
The changes in this release are as follows:
tv_grab_uk_rt has been tweaked to make fewer pages (including a
--limit-details option to get program details for only some times of
day) and to fix a couple of bugs. tv_grab_nl_wolf now skips over
empty data files at the source instead of dying. A new grabber
tv_grab_es_digital fetches listings for Spanish Digital+ channels.