New upstream release;
0.5.13
The changes in this release are as follows:
In lots of previous releases (VOB) data was copied from the wrong
position since the offset was fixed in the program whereas it is not
fixed on the DVD. This major bug in the -m and -O function is now
fixed. Hand-mounted DVD drives which are not listed in fstab are now
found.