New upstream release;
1.1.6
The changes in this release are as follows:
This release has a much improved jitter buffer, and some more tuning
to the denoiser and echo canceller. There is initial support (will
change in the future) for the Symbian OS. The speexenc/speexdec now
support gapless playback. The includes have been moved to
/usr/include/speex/, and there is now a run-time call for
identification of Speex version.