New upstream release;
1.0.2
The changes in this release are as follows:
The code now returns error on proxy authentication failure. It is now
also possible to set proxy credentials without storing them as clear
text in memory. A minor parsing bug was fixed.

1.0.1
The changes in this release are as follows:
This release fixes some compiler warnings for Solaris and Cygwin, and
a compilation problem on MacOS X. An option has been added to the
configure script to disable libcdio support.

1.0.0
The changes in this release are as follows:
Memory leak fixes were made in regular expression handling. An iconv
compilation fix was made for Solaris. Support for proxy
authentication was added.