New upstream PCB (20140316-1)
Tested on 10.7 and 10.9 with Xcode 4.6.3 and 5.1, respectively.

Note that this takes a while to start up on my 10.9 test machine. It seemed to
be competing with fc-cache for CPU, and it would just show a gray X11 window
for a bit. Subsequent startups were much faster.

Also, on 10.9, it does not seem to like X11-over-SSH. Might need to build a
non-OpenGL version if that becomes an issue.